add support sort spill

+++ core/src/expression/parser/parser.cpp
@@ -239,7 +239,7 @@ Expr *Parser::ParseRowExpressionHelper(string opStr, vector<Expr *> args)
     vector<DataTypeId> argTypes(args.size());
     std::transform(args.begin(), args.end(), argTypes.begin(),
         [](Expr *expr) -> DataTypeId { return expr->GetReturnTypeId(); });
-    for (int i = 0; i < argTypes.size(); i++) {
+    for (size_t i = 0; i < argTypes.size(); i++) {
         if (argTypes[i] == omniruntime::type::OMNI_DATE32) {
             argTypes[i] = omniruntime::type::OMNI_INT;
         }
